working on the DIS structure functions

+++ inc/apfel/distribution.h
@@ -37,7 +37,7 @@ namespace apfel
     // operators
     Distribution& operator  = (Distribution const& d); //!< this  = Distribution
     Distribution& operator *= (double const& s);       //!< this *= Scalar
-    Distribution& operator /= (int const& s);          //!< this /= Scalar
+    Distribution& operator /= (double const& s);       //!< this /= Scalar
     Distribution& operator *= (Distribution const& d); //!< this *= Distribution
     Distribution& operator += (Distribution const& d); //!< this += Distribution
     Distribution& operator -= (Distribution const& d); //!< this -= Distribution
@@ -46,7 +46,7 @@ namespace apfel
   // Extra operation definitions where Distribution is at the left hand side (lhs).
   Distribution operator * (double const& s, Distribution rhs);         //!< Scalar*Distribution
   Distribution operator * (Distribution lhs, double const& s);         //!< Distribution*Scalar
-  Distribution operator / (Distribution lhs, int const& s);            //!< Distribution/Scalar
+  Distribution operator / (Distribution lhs, double const& s);         //!< Distribution/Scalar
   Distribution operator + (Distribution lhs, Distribution const& rhs); //!< Distribution+Distribution
   Distribution operator - (Distribution lhs, Distribution const& rhs); //!< Distribution-Distribution
 }
